Changed the width of the input & button @input book
